requirements.txt
aiogram
pymysql
python-dotenv